requirements.txt
asgiref==3.5.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
defusedxml==0.7.1
Django==4.0.3
django-allauth==0.50.0
django-cors-headers==3.13.0
django-rest-knox==4.2.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-api-key==2.2.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
google-api-python-client==1.6.4
html5lib==1.1
httplib2==0.20.4
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
jsonpickle==0.9.6
MarkupSafe==2.1.1
more-itertools==8.12.0
oauth2client==4.1.2
oauthlib==3.2.0
Pillow==9.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.7
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
uritemplate==3.0.1
urllib3==1.26.9
webencodings==0.5.1
